requirements.txt
Flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0
dnspython==2.6.1
certifi==2024.2.2
